+++ torch/_dynamo/convert_frame.py
@@ -25,7 +25,6 @@ from __future__ import annotations
 
 import collections
 import contextlib
-import cProfile
 import dataclasses
 import dis
 import functools
@@ -35,11 +34,8 @@ import inspect
 import itertools
 import logging
 import os
-import pstats
 import random  # noqa: F401 -- eval_frame_cpp.cpp imports random at runtime; torch.package needs this to detect the dependency
-import subprocess
 import sys
-import tempfile
 import threading
 import time
 import traceback
@@ -48,7 +44,6 @@ import typing
 import unittest.mock as mock
 import weakref
 from dataclasses import dataclass
-from pathlib import Path
 from types import CellType, CodeType, FunctionType, ModuleType
 from typing import Any, cast, NoReturn, TypeVar
 from typing_extensions import ParamSpec
@@ -62,11 +57,7 @@ from torch._dynamo.distributed import get_compile_pg
 from torch._dynamo.symbolic_convert import TensorifyState
 from torch._guards import compile_context, CompileContext, CompileId, tracing
 from torch._logging import structured
-from torch._utils_internal import (
-    compile_time_strobelight_meta,
-    maybe_upload_prof_stats_to_manifold,
-    signpost_event,
-)
+from torch._utils_internal import compile_time_strobelight_meta, signpost_event
 from torch.fx._lazy_graph_module import _use_lazy_graph_module
 from torch.fx.experimental.symbolic_shapes import (
     ConstraintViolationError,
@@ -83,7 +74,14 @@ from torch.utils._python_dispatch import (
 )
 from torch.utils._traceback import CapturedTraceback, format_traceback_short
 
-from . import config, decorators, exc, graph_break_hints, trace_rules
+from . import (
+    config,
+    decorators,
+    exc,
+    graph_break_hints,
+    profiler as dynamo_profiler,
+    trace_rules,
+)
 from .backends.registry import _is_registered_backend
 from .bytecode_analysis import remove_dead_code, remove_pointless_jumps
 from .bytecode_transformation import (
@@ -200,6 +198,10 @@ if typing.TYPE_CHECKING:
 log = logging.getLogger(__name__)
 bytecode_log = torch._logging.getArtifactLogger(__name__, "bytecode")
 graph_break_log = torch._logging.getArtifactLogger(__name__, "graph_breaks")
+
+
+cprofile_wrapper = dynamo_profiler.cprofile_wrapper
+maybe_cprofile = dynamo_profiler.maybe_cprofile
 
 
 compile_lock = threading.RLock()
@@ -485,88 +487,6 @@ def exception_handler(
 
 FRAME_COUNTER = 0
 FRAME_COMPILE_COUNTER: typing.Counter[int | FrameStateSizeEntry] = collections.Counter()
-
-
-def maybe_cprofile(func: Callable[_P, _T]) -> Callable[_P, _T]:
-    if config.cprofile:
-        return cprofile_wrapper(func)
-    return func
-
-
-def cprofile_wrapper(func: Callable[_P, _T]) -> Callable[_P, _T]:
-    @functools.wraps(func)
-    def profile_wrapper(*args: _P.args, **kwargs: _P.kwargs) -> _T:
-        trace_id = CompileContext.current_trace_id()
-        if not trace_id:
-            raise AssertionError("Trace id is None")
-        profile_path = Path(
-            os.path.join(
-                tempfile.gettempdir(),
-                f"{func.__name__}_{str(trace_id).replace('/', '_')}.profile",
-            )
-        )
-        prof = cProfile.Profile()
-        try:
-            start_ts = time.time()
-            # runcall calls prof.enable() and prof.disable(), so do NOT call
-            # enable outside. This leads to issues like
-            # ValueError: Another profiling tool is already active
-            # pyrefly: ignore [bad-argument-type]
-            retval = prof.runcall(func, *args, **kwargs)
-            profile_latency = time.time() - start_ts
-        except ValueError:
-            log.exception("failed to enable cProfile")
-            profile_latency = 0
-            retval = func(*args, **kwargs)
-        log.warning(
-            "### Cprofile for %s trace id [%s] took %.3f seconds ###",
-            func.__name__,
-            trace_id,
-            profile_latency,
-        )
-        ps = pstats.Stats(prof)
-        try:
-            prof.dump_stats(profile_path)
-        except OSError:
-            log.exception("Cannot write to %s", profile_path)
-        log.warning("Raw profile at %s", profile_path)
-        svg_path = profile_path.with_suffix(".svg")
-        try:
-            with subprocess.Popen(
-                [
-                    "gprof2dot",
-                    "-f",
-                    "pstats",
-                    "--node-label=total-time-percentage",
-                    "--node-label=self-time-percentage",
-                    "--node-label=total-time",
-                    str(profile_path),
-                ],
-                stdout=subprocess.PIPE,
-            ) as gprof2dot_process:
-                subprocess.check_call(
-                    ["dot", "-Tsvg", "-o", str(svg_path)],
-                    stdin=gprof2dot_process.stdout,
-                )
-                log.warning("Generated SVG from profile at %s", svg_path)
-        except FileNotFoundError:
-            log.warning(
-                "Failed to generate SVG from profile -- dumping stats instead."
-                "Try installing gprof2dot and dot for a better visualization"
-            )
-            ps.sort_stats(pstats.SortKey.TIME).print_stats(20)
-            ps.sort_stats(pstats.SortKey.CUMULATIVE).print_stats(20)
-
-        if manifold_link := maybe_upload_prof_stats_to_manifold(
-            str(profile_path)
-        ):  # fb-only
-            torch._logging.trace_structured(
-                "link",
-                lambda: {"name": "cprofile_manifold_url", "url": manifold_link},
-            )
-        return retval
-
-    return profile_wrapper
 
 
 @dataclass
